requirements.txt
torch==1.9.0+cu111
torch-tb-profiler==0.4.0
torchaudio==0.9.0
torchmetrics==0.9.0
torchvision==0.10.0+cu111
PyYAML==6.0
matplotlib==3.1.2
matplotlib-inline==0.1.3
numpy==1.22.4
pandas==1.4.2
bluepy==1.3.0